refactor(mcp): one file per MCP tool
server.go grew to 1127 LOC with seven tools' input/output structs +
handlers + serverInstructions all mixed together. Splits each tool
into its own file: search_tool.go, read_attributes_tool.go,
read_lines_tool.go, find_duplicates_tool.go, stats_tool.go,
index_stats_tool.go, list_attributes_tool.go. server.go shrinks to
the wiring only — handlers struct, instructions text, New + Run.
Pure mechanical refactor: no behaviour, no public API, no test
changes.
