+++ internal/mcpserver/list_attributes_tool.go
@@ -1,0 +1,34 @@
+package mcpserver
+
+import (
+	"context"
+
+	"github.com/modelcontextprotocol/go-sdk/mcp"
+
+	"github.com/richardwooding/file-search-on/internal/celexpr"
+	"github.com/richardwooding/file-search-on/internal/content"
+)
+
+// ContentTypeDoc describes a registered content type.
+type ContentTypeDoc struct {
+	Name       string   `json:"name"`
+	Extensions []string `json:"extensions"`
+}
+
+// ListAttributesOutput is the structured output of the `list_attributes` tool.
+type ListAttributesOutput struct {
+	Schema       celexpr.SchemaDoc `json:"schema"`
+	ContentTypes []ContentTypeDoc  `json:"content_types"`
+}
+
+func listAttributesHandler(_ context.Context, _ *mcp.CallToolRequest, _ struct{}) (*mcp.CallToolResult, ListAttributesOutput, error) {
+	types := content.DefaultRegistry().Types()
+	docs := make([]ContentTypeDoc, len(types))
+	for i, t := range types {
+		docs[i] = ContentTypeDoc{Name: t.Name(), Extensions: t.Extensions()}
+	}
+	return nil, ListAttributesOutput{
+		Schema:       celexpr.Schema(),
+		ContentTypes: docs,
+	}, nil
+}
