+++ internal/mcpserver/search_tool.go
@@ -1,0 +1,590 @@
+package mcpserver
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"sort"
+	"time"
+
+	"github.com/modelcontextprotocol/go-sdk/mcp"
+
+	"github.com/richardwooding/file-search-on/internal/content"
+	"github.com/richardwooding/file-search-on/internal/search"
+)
+
+// SearchInput is the JSON-schema input for the `search` tool.
+type SearchInput struct {
+	Expr             string   `json:"expr,omitempty" jsonschema:"CEL expression matched against file attributes. Boolean type predicates: is_markdown, is_pdf, is_html, is_xml, is_json, is_csv, is_text, is_image, is_audio, is_video, is_office, is_epub, is_archive, is_binary, is_email, is_source. Common attributes: size (int, bytes), name/path/dir/ext (string), word_count/line_count/page_count (int), title/author/language (string). Examples: 'is_markdown && word_count > 500'; 'is_pdf && page_count > 10'; 'is_image && iso > 1600'; 'is_audio && sample_rate >= 96000'; 'is_video && duration > 1800'; 'is_source && language == \"go\" && loc > 100'; 'size > 1000000 && !is_binary'. Empty means match all. Call list_attributes for the full schema."`
+	Dir              string   `json:"dir,omitempty" jsonschema:"Directory to search in. Defaults to '.'. Ignored when 'dirs' is non-empty."`
+	Dirs             []string `json:"dirs,omitempty" jsonschema:"Multiple directories to search in one call. When non-empty, takes precedence over 'dir'. Each root's .gitignore is honoured independently when respect_gitignore is set."`
+	Workers          int      `json:"workers,omitempty" jsonschema:"Number of parallel workers. Defaults to runtime.NumCPU()."`
+	MaxLineBytes     int      `json:"max_line_bytes,omitempty" jsonschema:"Per-line scanner buffer cap for text/CSV/HTML (bytes). 0 uses the 1 MiB default; raise for very long log lines."`
+	TimeoutSeconds   *float64 `json:"timeout_seconds,omitempty" jsonschema:"Override the server's default per-call timeout for this invocation (in seconds; fractions allowed). Omit to use the server default (set when the MCP server was started). Pass 0 to disable the timeout for this call. On expiry the walk is cancelled and the partial result set is returned with cancelled=true."`
+	SortBy           string   `json:"sort_by,omitempty" jsonschema:"Sort matches by attribute. Recognised keys: size, name, path, mod_time, word_count, line_count, page_count, duration, bitrate, sample_rate, video_height, video_width, frame_rate, iso, focal_length, taken_at, sent_at, year, entry_count, uncompressed_size, loc, attachment_count, email_count. Files missing the attribute group at the end. Sorting buffers the full result set (top-K is incoherent with streaming)."`
+	Order            string   `json:"order,omitempty" jsonschema:"Sort direction when sort_by is set: 'asc' (default) or 'desc'."`
+	Limit            int      `json:"limit,omitempty" jsonschema:"Cap the returned match count. With sort_by, returns the top-N (after sorting). Without sort_by, returns the first N in walk order. 0 = unlimited."`
+	IncludeSnippet   bool     `json:"include_snippet,omitempty" jsonschema:"When true, populate each match's 'snippet' field with the first N lines of the file body (see snippet_lines). Only text-based content types (markdown / text / html / csv / json / xml / source/*) populate; binary families leave snippet empty."`
+	SnippetLines     int      `json:"snippet_lines,omitempty" jsonschema:"How many lines to include per snippet (default 10). Ignored when include_snippet is false."`
+	IncludeBody      bool     `json:"include_body,omitempty" jsonschema:"When true, the full file body is exposed to the CEL expression as the 'body' string variable, so filters like body.contains(\"transformer\") or body.matches(\"\\\\bAPI\\\\b\") run at search time. Only text-based content types populate body; capped at body_max_bytes (default 1 MiB). Expensive: reads every candidate file's body, not just headers — pair with tight expr / excludes / timeout."`
+	BodyMaxBytes     int      `json:"body_max_bytes,omitempty" jsonschema:"Cap on the body string in bytes (default 1 MiB). Files larger than the cap are silently truncated; the prefix still participates in body.contains / body.matches. Ignored when include_body is false."`
+	Excludes         []string `json:"excludes,omitempty" jsonschema:"Glob patterns matched against the basename of each file/directory; matched directories are pruned. Example: ['node_modules', '.git', 'target', '*.bak']. Use respect_gitignore for path-aware patterns."`
+	RespectGitignore bool     `json:"respect_gitignore,omitempty" jsonschema:"When true, parse a .gitignore at the walk root (if present) and skip matching paths. Honours standard gitignore semantics. Nested .gitignore files in subdirectories are NOT honoured in this version."`
+}
+
+// SearchMatch is one match returned by the `search` tool. Beyond path /
+// content_type / size, every CEL-visible attribute is included when the
+// matched content type emits it; absent fields are omitted from the JSON
+// payload so simple consumers see a compact shape.
+type SearchMatch struct {
+	Path        string `json:"path"`
+	ContentType string `json:"content_type"`
+	Size        int64  `json:"size"`
+
+	Title    string `json:"title,omitempty"`
+	Author   string `json:"author,omitempty"`
+	Language string `json:"language,omitempty"`
+
+	WordCount   int64 `json:"word_count,omitempty"`
+	LineCount   int64 `json:"line_count,omitempty"`
+	PageCount   int64 `json:"page_count,omitempty"`
+	ColumnCount int64 `json:"column_count,omitempty"`
+
+	CSVColumns  []string `json:"csv_columns,omitempty"`
+	RootElement string   `json:"root_element,omitempty"`
+	JSONKind    string   `json:"json_kind,omitempty"`
+
+	ImgWidth  int64 `json:"img_width,omitempty"`
+	ImgHeight int64 `json:"img_height,omitempty"`
+
+	CameraMake   string  `json:"camera_make,omitempty"`
+	CameraModel  string  `json:"camera_model,omitempty"`
+	Lens         string  `json:"lens,omitempty"`
+	TakenAt      string  `json:"taken_at,omitempty"` // RFC3339 when set
+	Orientation  int64   `json:"orientation,omitempty"`
+	GPSLat       float64 `json:"gps_lat,omitempty"`
+	GPSLon       float64 `json:"gps_lon,omitempty"`
+	ISO          int64   `json:"iso,omitempty"`
+	FocalLength  float64 `json:"focal_length,omitempty"`
+	FStop        float64 `json:"f_stop,omitempty"`
+	ExposureTime float64 `json:"exposure_time,omitempty"`
+
+	FrontmatterFormat string         `json:"frontmatter_format,omitempty"`
+	Frontmatter       map[string]any `json:"frontmatter,omitempty"`
+	Tags              []string       `json:"tags,omitempty"`
+	Categories        []string       `json:"categories,omitempty"`
+	Draft             bool           `json:"draft,omitempty"`
+	Date              string         `json:"date,omitempty"` // RFC3339 when set
+
+	IsMarkdown bool `json:"is_markdown,omitempty"`
+	IsJSON     bool `json:"is_json,omitempty"`
+	IsXML      bool `json:"is_xml,omitempty"`
+	IsHTML     bool `json:"is_html,omitempty"`
+	IsPDF      bool `json:"is_pdf,omitempty"`
+	IsImage    bool `json:"is_image,omitempty"`
+	IsText     bool `json:"is_text,omitempty"`
+	IsCSV      bool `json:"is_csv,omitempty"`
+	IsEPUB     bool `json:"is_epub,omitempty"`
+	IsOffice   bool `json:"is_office,omitempty"`
+	IsAudio    bool `json:"is_audio,omitempty"`
+	IsVideo    bool `json:"is_video,omitempty"`
+	IsArchive  bool `json:"is_archive,omitempty"`
+	IsBinary   bool `json:"is_binary,omitempty"`
+	IsEmail    bool `json:"is_email,omitempty"`
+	IsSource   bool `json:"is_source,omitempty"`
+	IsNotebook bool `json:"is_notebook,omitempty"`
+
+	Artist      string `json:"artist,omitempty"`
+	Album       string `json:"album,omitempty"`
+	AlbumArtist string `json:"album_artist,omitempty"`
+	Composer    string `json:"composer,omitempty"`
+	Year        int64  `json:"year,omitempty"`
+	Track       int64  `json:"track,omitempty"`
+	Genre       string `json:"genre,omitempty"`
+
+	Duration       float64 `json:"duration,omitempty"`
+	Bitrate        int64   `json:"bitrate,omitempty"`
+	NominalBitrate int64   `json:"nominal_bitrate,omitempty"`
+	SampleRate     int64   `json:"sample_rate,omitempty"`
+	Channels       int64   `json:"channels,omitempty"`
+	BitDepth       int64   `json:"bit_depth,omitempty"`
+
+	VideoCodec  string  `json:"video_codec,omitempty"`
+	AudioCodec  string  `json:"audio_codec,omitempty"`
+	VideoWidth  int64   `json:"video_width,omitempty"`
+	VideoHeight int64   `json:"video_height,omitempty"`
+	FrameRate   float64 `json:"frame_rate,omitempty"`
+	Rotation    int64   `json:"rotation,omitempty"`
+
+	ColorPrimaries string `json:"color_primaries,omitempty"`
+	ColorTransfer  string `json:"color_transfer,omitempty"`
+	IsHDR          bool   `json:"is_hdr,omitempty"`
+
+	Subtitles         bool     `json:"subtitles,omitempty"`
+	SubtitleLanguages []string `json:"subtitle_languages,omitempty"`
+
+	ReplayGainTrackGain float64 `json:"replaygain_track_gain,omitempty"`
+	ReplayGainAlbumGain float64 `json:"replaygain_album_gain,omitempty"`
+
+	EntryCount       int64    `json:"entry_count,omitempty"`
+	UncompressedSize int64    `json:"uncompressed_size,omitempty"`
+	TopLevelEntries  []string `json:"top_level_entries,omitempty"`
+	HasRootDir       bool     `json:"has_root_dir,omitempty"`
+
+	Architectures       []string `json:"architectures,omitempty"`
+	Bitness             int64    `json:"bitness,omitempty"`
+	BinaryFormat        string   `json:"binary_format,omitempty"`
+	BinaryType          string   `json:"binary_type,omitempty"`
+	IsDynamicallyLinked bool     `json:"is_dynamically_linked,omitempty"`
+	IsStripped          bool     `json:"is_stripped,omitempty"`
+	EntryPoint          int64    `json:"entry_point,omitempty"`
+
+	EmailTo         []string `json:"email_to,omitempty"`
+	EmailCc         []string `json:"email_cc,omitempty"`
+	EmailMessageID  string   `json:"email_message_id,omitempty"`
+	EmailInReplyTo  string   `json:"email_in_reply_to,omitempty"`
+	SentAt          string   `json:"sent_at,omitempty"` // RFC3339 when set
+	AttachmentCount int64    `json:"attachment_count,omitempty"`
+	EmailCount      int64    `json:"email_count,omitempty"`
+
+	LOC        int64 `json:"loc,omitempty"`
+	CommentLOC int64 `json:"comment_loc,omitempty"`
+	BlankLOC   int64 `json:"blank_loc,omitempty"`
+
+	CellCount         int64  `json:"cell_count,omitempty"`
+	CodeCellCount     int64  `json:"code_cell_count,omitempty"`
+	MarkdownCellCount int64  `json:"markdown_cell_count,omitempty"`
+	Kernel            string `json:"kernel,omitempty"`
+
+	// Snippet is the first N lines of the file body when the search
+	// call had include_snippet=true and the content type is
+	// text-based. Empty otherwise. Lets an agent decide whether a
+	// match is relevant without a separate read_attributes round trip.
+	Snippet string `json:"snippet,omitempty"`
+}
+
+// matchFrom projects a search.Result (with Attrs populated) into a
+// SearchMatch wire object.
+func matchFrom(r search.Result) SearchMatch {
+	m := SearchMatch{
+		Path:        r.Path,
+		ContentType: r.ContentType,
+		Size:        r.Size,
+		Snippet:     r.Snippet,
+	}
+	if r.Attrs == nil {
+		return m
+	}
+	a := r.Attrs
+	m.IsMarkdown, m.IsJSON, m.IsXML, m.IsHTML = a.IsMarkdown, a.IsJSON, a.IsXML, a.IsHTML
+	m.IsPDF, m.IsImage = a.IsPDF, a.IsImage
+	m.IsText, m.IsCSV, m.IsEPUB, m.IsOffice = a.IsText, a.IsCSV, a.IsEPUB, a.IsOffice
+	m.IsAudio = a.IsAudio
+	m.IsVideo = a.IsVideo
+	m.IsArchive = a.IsArchive
+	m.IsBinary = a.IsBinary
+	m.IsEmail = a.IsEmail
+	m.IsSource = a.IsSource
+	m.IsNotebook = a.IsNotebook
+
+	if a.Extra == nil {
+		return m
+	}
+	if v, ok := a.Extra["title"].(string); ok {
+		m.Title = v
+	}
+	if v, ok := a.Extra["author"].(string); ok {
+		m.Author = v
+	}
+	if v, ok := a.Extra["language"].(string); ok {
+		m.Language = v
+	}
+	if v, ok := a.Extra["word_count"].(int64); ok {
+		m.WordCount = v
+	}
+	if v, ok := a.Extra["line_count"].(int64); ok {
+		m.LineCount = v
+	}
+	if v, ok := a.Extra["page_count"].(int64); ok {
+		m.PageCount = v
+	}
+	if v, ok := a.Extra["column_count"].(int64); ok {
+		m.ColumnCount = v
+	}
+	if v, ok := a.Extra["csv_columns"].([]string); ok {
+		m.CSVColumns = v
+	}
+	if v, ok := a.Extra["root_element"].(string); ok {
+		m.RootElement = v
+	}
+	if v, ok := a.Extra["kind"].(string); ok {
+		m.JSONKind = v
+	}
+	if v, ok := a.Extra["width"].(int64); ok {
+		m.ImgWidth = v
+	}
+	if v, ok := a.Extra["height"].(int64); ok {
+		m.ImgHeight = v
+	}
+	if v, ok := a.Extra["camera_make"].(string); ok {
+		m.CameraMake = v
+	}
+	if v, ok := a.Extra["camera_model"].(string); ok {
+		m.CameraModel = v
+	}
+	if v, ok := a.Extra["lens"].(string); ok {
+		m.Lens = v
+	}
+	if v, ok := a.Extra["taken_at"].(time.Time); ok && !v.IsZero() {
+		m.TakenAt = v.Format(time.RFC3339)
+	}
+	if v, ok := a.Extra["orientation"].(int64); ok {
+		m.Orientation = v
+	}
+	if v, ok := a.Extra["gps_lat"].(float64); ok {
+		m.GPSLat = v
+	}
+	if v, ok := a.Extra["gps_lon"].(float64); ok {
+		m.GPSLon = v
+	}
+	if v, ok := a.Extra["iso"].(int64); ok {
+		m.ISO = v
+	}
+	if v, ok := a.Extra["focal_length"].(float64); ok {
+		m.FocalLength = v
+	}
+	if v, ok := a.Extra["f_stop"].(float64); ok {
+		m.FStop = v
+	}
+	if v, ok := a.Extra["exposure_time"].(float64); ok {
+		m.ExposureTime = v
+	}
+	if v, ok := a.Extra["artist"].(string); ok {
+		m.Artist = v
+	}
+	if v, ok := a.Extra["album"].(string); ok {
+		m.Album = v
+	}
+	if v, ok := a.Extra["album_artist"].(string); ok {
+		m.AlbumArtist = v
+	}
+	if v, ok := a.Extra["composer"].(string); ok {
+		m.Composer = v
+	}
+	if v, ok := a.Extra["year"].(int64); ok {
+		m.Year = v
+	}
+	if v, ok := a.Extra["track"].(int64); ok {
+		m.Track = v
+	}
+	if v, ok := a.Extra["genre"].(string); ok {
+		m.Genre = v
+	}
+	if v, ok := a.Extra["duration"].(float64); ok {
+		m.Duration = v
+	}
+	if v, ok := a.Extra["bitrate"].(int64); ok {
+		m.Bitrate = v
+	}
+	if v, ok := a.Extra["sample_rate"].(int64); ok {
+		m.SampleRate = v
+	}
+	if v, ok := a.Extra["channels"].(int64); ok {
+		m.Channels = v
+	}
+	if v, ok := a.Extra["bit_depth"].(int64); ok {
+		m.BitDepth = v
+	}
+	if v, ok := a.Extra["nominal_bitrate"].(int64); ok {
+		m.NominalBitrate = v
+	}
+	if v, ok := a.Extra["video_codec"].(string); ok {
+		m.VideoCodec = v
+	}
+	if v, ok := a.Extra["audio_codec"].(string); ok {
+		m.AudioCodec = v
+	}
+	if v, ok := a.Extra["video_width"].(int64); ok {
+		m.VideoWidth = v
+	}
+	if v, ok := a.Extra["video_height"].(int64); ok {
+		m.VideoHeight = v
+	}
+	if v, ok := a.Extra["frame_rate"].(float64); ok {
+		m.FrameRate = v
+	}
+	if v, ok := a.Extra["rotation"].(int64); ok {
+		m.Rotation = v
+	}
+	if v, ok := a.Extra["color_primaries"].(string); ok {
+		m.ColorPrimaries = v
+	}
+	if v, ok := a.Extra["color_transfer"].(string); ok {
+		m.ColorTransfer = v
+	}
+	if v, ok := a.Extra["is_hdr"].(bool); ok {
+		m.IsHDR = v
+	}
+	if v, ok := a.Extra["subtitles"].(bool); ok {
+		m.Subtitles = v
+	}
+	if v, ok := a.Extra["subtitle_languages"].([]string); ok && len(v) > 0 {
+		m.SubtitleLanguages = v
+	}
+	if v, ok := a.Extra["replaygain_track_gain"].(float64); ok {
+		m.ReplayGainTrackGain = v
+	}
+	if v, ok := a.Extra["replaygain_album_gain"].(float64); ok {
+		m.ReplayGainAlbumGain = v
+	}
+	if v, ok := a.Extra["entry_count"].(int64); ok {
+		m.EntryCount = v
+	}
+	if v, ok := a.Extra["uncompressed_size"].(int64); ok {
+		m.UncompressedSize = v
+	}
+	if v, ok := a.Extra["top_level_entries"].([]string); ok && len(v) > 0 {
+		m.TopLevelEntries = v
+	}
+	if v, ok := a.Extra["has_root_dir"].(bool); ok {
+		m.HasRootDir = v
+	}
+	if v, ok := a.Extra["architectures"].([]string); ok && len(v) > 0 {
+		m.Architectures = v
+	}
+	if v, ok := a.Extra["bitness"].(int64); ok {
+		m.Bitness = v
+	}
+	if v, ok := a.Extra["binary_format"].(string); ok {
+		m.BinaryFormat = v
+	}
+	if v, ok := a.Extra["binary_type"].(string); ok {
+		m.BinaryType = v
+	}
+	if v, ok := a.Extra["is_dynamically_linked"].(bool); ok {
+		m.IsDynamicallyLinked = v
+	}
+	if v, ok := a.Extra["is_stripped"].(bool); ok {
+		m.IsStripped = v
+	}
+	if v, ok := a.Extra["entry_point"].(int64); ok {
+		m.EntryPoint = v
+	}
+	if v, ok := a.Extra["email_to"].([]string); ok && len(v) > 0 {
+		m.EmailTo = v
+	}
+	if v, ok := a.Extra["email_cc"].([]string); ok && len(v) > 0 {
+		m.EmailCc = v
+	}
+	if v, ok := a.Extra["email_message_id"].(string); ok {
+		m.EmailMessageID = v
+	}
+	if v, ok := a.Extra["email_in_reply_to"].(string); ok {
+		m.EmailInReplyTo = v
+	}
+	if v, ok := a.Extra["sent_at"].(time.Time); ok && !v.IsZero() {
+		m.SentAt = v.Format(time.RFC3339)
+	}
+	if v, ok := a.Extra["attachment_count"].(int64); ok {
+		m.AttachmentCount = v
+	}
+	if v, ok := a.Extra["email_count"].(int64); ok {
+		m.EmailCount = v
+	}
+	if v, ok := a.Extra["loc"].(int64); ok {
+		m.LOC = v
+	}
+	if v, ok := a.Extra["comment_loc"].(int64); ok {
+		m.CommentLOC = v
+	}
+	if v, ok := a.Extra["blank_loc"].(int64); ok {
+		m.BlankLOC = v
+	}
+	if v, ok := a.Extra["cell_count"].(int64); ok {
+		m.CellCount = v
+	}
+	if v, ok := a.Extra["code_cell_count"].(int64); ok {
+		m.CodeCellCount = v
+	}
+	if v, ok := a.Extra["markdown_cell_count"].(int64); ok {
+		m.MarkdownCellCount = v
+	}
+	if v, ok := a.Extra["kernel"].(string); ok {
+		m.Kernel = v
+	}
+	if v, ok := a.Extra["frontmatter_format"].(string); ok {
+		m.FrontmatterFormat = v
+	}
+	if v, ok := a.Extra["frontmatter"].(map[string]any); ok && len(v) > 0 {
+		m.Frontmatter = v
+	}
+	if v, ok := a.Extra["tags"].([]string); ok && len(v) > 0 {
+		m.Tags = v
+	}
+	if v, ok := a.Extra["categories"].([]string); ok && len(v) > 0 {
+		m.Categories = v
+	}
+	if v, ok := a.Extra["draft"].(bool); ok {
+		m.Draft = v
+	}
+	if v, ok := a.Extra["date"].(time.Time); ok && !v.IsZero() {
+		m.Date = v.Format(time.RFC3339)
+	}
+	return m
+}
+
+// SearchOutput is the structured output of the `search` tool.
+//
+// When Cancelled is true, the walk did not complete; Matches contains
+// every result that was emitted by the walker before the deadline /
+// signal fired. CancellationReason distinguishes "timeout" (server
+// default or per-call timeout_seconds expired) from "client_cancel"
+// (the MCP client closed the request or the parent ctx was cancelled
+// for some other reason). ElapsedSeconds reports wall-clock time spent
+// inside the search handler — useful for tuning timeouts.
+type SearchOutput struct {
+	Matches            []SearchMatch `json:"matches"`
+	Count              int           `json:"count"`
+	Cancelled          bool          `json:"cancelled,omitempty"`
+	CancellationReason string        `json:"cancellation_reason,omitempty"`
+	ElapsedSeconds     float64       `json:"elapsed_seconds,omitempty"`
+}
+
+// progressNotifyStride is the number of matches between two
+// notifications/progress messages. Smaller searches (< stride matches)
+// emit zero notifications and just land in one final response. Tunable
+// later via an Options field if a client needs finer granularity.
+const progressNotifyStride = 50
+
+func (h *handlers) searchHandler(ctx context.Context, req *mcp.CallToolRequest, in SearchInput) (*mcp.CallToolResult, SearchOutput, error) {
+	expr := in.Expr
+	if expr == "" {
+		expr = "true"
+	}
+	dir := in.Dir
+	if dir == "" {
+		dir = "."
+	}
+
+	// Resolve the effective timeout: per-call override > server default.
+	// A nil pointer inherits the default; an explicit 0 means "no
+	// timeout for this call" (the parent ctx still applies). We track
+	// the parent ctx separately so we can distinguish a server-level
+	// cancellation (transport close, parent ctx) from our own
+	// timeout firing.
+	parentCtx := ctx
+	timeout := h.defaultTimeout
+	if in.TimeoutSeconds != nil {
+		timeout = time.Duration(*in.TimeoutSeconds * float64(time.Second))
+	}
+	if timeout > 0 {
+		var cancel context.CancelFunc
+		ctx, cancel = context.WithTimeout(ctx, timeout)
+		defer cancel()
+	}
+
+	start := time.Now()
+
+	out := make(chan search.Result, 64)
+	var walkErr error
+	done := make(chan struct{})
+	// The MCP handler always buffers results (it sorts by path
+	// before returning) so we route sort/limit through search.Walk
+	// rather than re-implementing the post-sort here. But progress
+	// notifications + cancellation handling still want streaming —
+	// so we feed the channel ourselves and sort/limit the collected
+	// matches post-stream using the same sortAndLimit helper.
+	// Multi-dir: in.Dirs wins when non-empty; else fall back to
+	// the single 'dir' field (with default "." applied above).
+	walkOpts := search.Options{
+		Root:              dir,
+		Roots:             in.Dirs,
+		Expr:              expr,
+		Workers:           in.Workers,
+		MaxLineBytes:      in.MaxLineBytes,
+		IncludeAttributes: true,
+		Index:             h.idx,
+		SnippetLines:      in.SnippetLines,
+		IncludeSnippet:    in.IncludeSnippet,
+		IncludeBody:       in.IncludeBody,
+		BodyMaxBytes:      in.BodyMaxBytes,
+		Excludes:          in.Excludes,
+		RespectGitignore:  in.RespectGitignore,
+		// Sort, Order, Limit are applied via sortAndLimit AFTER we
+		// collect — see end of handler. We don't pass them to
+		// WalkStream because WalkStream doesn't honour them.
+	}
+	go func() {
+		walkErr = search.WalkStream(ctx, walkOpts, content.DefaultRegistry(), out)
+		close(done)
+	}()
+
+	// Drain the channel as matches arrive. Emit a progress notification
+	// every `progressNotifyStride` matches when the client passed a
+	// progressToken — the SDK's NotifyProgress is a no-op for clients
+	// that didn't request progress.
+	//
+	// We collect raw search.Results here (not the projected
+	// SearchMatch wire shape) so sort_by has access to the full
+	// FileAttributes for per-family scalar keys. Projection happens
+	// after the sort.
+	token := req.Params.GetProgressToken()
+	var results []search.Result
+	processed := int64(0)
+	for r := range out {
+		results = append(results, r)
+		processed++
+		if token != nil && processed%progressNotifyStride == 0 {
+			_ = req.Session.NotifyProgress(ctx, &mcp.ProgressNotificationParams{
+				ProgressToken: token,
+				Progress:      float64(processed),
+				Message:       fmt.Sprintf("%d matches so far", processed),
+			})
+		}
+	}
+	<-done
+
+	elapsed := time.Since(start).Seconds()
+
+	cancelled := errors.Is(walkErr, context.Canceled) || errors.Is(walkErr, context.DeadlineExceeded)
+	if walkErr != nil && !cancelled {
+		return nil, SearchOutput{}, fmt.Errorf("walk: %w", walkErr)
+	}
+
+	// Order: explicit sort_by > legacy path-sort default. Limit is
+	// applied AFTER sorting so the response respects top-K
+	// semantics. sortAndLimit lives in internal/search next to the
+	// type-aware compareByKey so this stays the single source of
+	// truth for sort/limit logic.
+	if in.SortBy != "" || in.Limit > 0 {
+		sortOpts := search.Options{Sort: in.SortBy, Order: in.Order, Limit: in.Limit}
+		results = search.SortAndLimit(results, sortOpts)
+	} else {
+		// Historical contract: matches sorted by path. Preserve it
+		// when the caller didn't request a specific sort.
+		sort.Slice(results, func(i, j int) bool { return results[i].Path < results[j].Path })
+	}
+
+	matches := make([]SearchMatch, len(results))
+	for i, r := range results {
+		matches[i] = matchFrom(r)
+	}
+
+	output := SearchOutput{
+		Matches:        matches,
+		Count:          len(matches),
+		ElapsedSeconds: elapsed,
+	}
+	if cancelled {
+		output.Cancelled = true
+		// "timeout" when our deadline fired and the parent ctx is
+		// still healthy; otherwise the parent (transport / client /
+		// process signal) is the cause.
+		if errors.Is(walkErr, context.DeadlineExceeded) && parentCtx.Err() == nil {
+			output.CancellationReason = "timeout"
+		} else {
+			output.CancellationReason = "client_cancel"
+		}
+	}
+	return nil, output, nil
+}
